fixing bugs

+++ users/models.py
@@ -33,7 +33,7 @@ class CustomUser(AbstractBaseUser, PermissionsMixin):
     )
     phone_number = models.CharField(max_length=15, unique=True)
     name = models.CharField(max_length=255, blank=True, null=True)
-    solde = models.FloatField(default=1000)
+    solde = models.DecimalField(max_digits=10, decimal_places=2,default=1000)
     is_active = models.BooleanField(default=True)
     is_staff = models.BooleanField(default=False)
     created_at = models.DateTimeField(auto_now_add=True)
